[style]: renamed the entities -> entity package
